# specs/001-md/contracts/api.yaml
openapi: 3.0.3
info:
  title: MOF数据t-SNE交互式可视化 API
  description: 材料科学研究用MOF数据t-SNE降维和可视化Web应用API
  version: 1.0.0
  contact:
    name: MOF Visualization Team
    email: team@mof-viz.com
servers:
  - url: http://localhost:8000
    description: 本地开发服务器
paths:
...
  /api/pipelines/{pipeline_id}/status:
    get:
      summary: 获取处理状态
      description: 获取指定处理流水线的实时状态和进度信息
      operationId: get_pipeline_status
      parameters:
        - name: pipeline_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 处理状态信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineStatus'
        '404':
          $ref: '#/components/responses/NotFound'

  # 可视化API
...
  /ws/pipelines/{pipeline_id}:
    get:
      summary: WebSocket连接
      description: 建立WebSocket连接以接收实时进度更新
      operationId: websocket_connection
      parameters:
        - name: pipeline_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
components:
  schemas:
...
    BaseResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 请求是否成功
        timestamp:
          type: string
          format: date-time
          description: 响应时间
        request_id:
          type: string
          format: uuid
          description: 请求唯一标识符
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            error:
              type: object
              properties:
                code:
                  type: string
                  description: 错误代码
                message:
                  type: string
                  description: 错误信息
                details:
                  type: object
                  description: 错误详情

    # 数据相关Schema
...
    PipelineStatus:
      type: object
      properties:
        pipeline_id:
          type: string
          format: uuid
        dataset_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, completed, failed]
        progress_percentage:
          type: number
          format: float
          minimum: 0
          maximum: 100
        current_step:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        total_duration_ms:
          type: integer
        steps:
          type: array
          items:
            $ref: '#/components/schemas/StepStatus'
        error_message:
          type: string
        performance_metrics:
          $ref: '#/components/schemas/PerformanceMetrics'
    StepStatus:
      type: object
      properties:
        step_name:
          type: string
        step_type:
          type: string
          enum: [preprocessing, pca, tsne, visualization]
        status:
          type: string
          enum: [pending, running, completed, failed]
        progress_percentage:
          type: number
          format: float
          minimum: 0
          maximum: 100
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        duration_ms:
          type: integer
        input_rows:
          type: integer
        output_rows:
          type: integer
        memory_usage_mb:
          type: number
          format: float
        error_message:
          type: string
    PerformanceMetrics:
      type: object
      properties:
        total_memory_mb:
          type: number
          format: float
        peak_memory_mb:
          type: number
          format: float
        total_cpu_time_ms:
          type: integer
        average_cpu_usage:
          type: number
          format: float
        disk_io_mb:
          type: number
          format: float

    # 可视化相关Schema
...
  responses:
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  error:
                    properties:
                      code:
                        example: NOT_FOUND
                      message:
                        example: 请求的资源不存在
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
...